+++ src/interface/new-views/dashboard.py
@@ -9,7 +9,6 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../.
 from services.search import search
 from services.graph import plotar_grafico_periodo, calcular_percentual_aumento_por_periodo
 from alertas import alertas_page
-from app import landing_page
 
 current_dir = Path(__file__).parent
 img_path = current_dir / "assets" / "img" / "Icon.png"
@@ -109,8 +108,6 @@ with st.sidebar:
     if st.button("Alertas"):
         change_page("Alertas")
     
-    if st.button("Início"):
-        change_page("Inicio")
 
 
     if st.button("Dashboard"):
@@ -211,7 +208,6 @@ elif st.session_state.current_page == "Alertas":
 elif st.session_state.current_page == "Alertas":
     alertas_page()
 
-elif st.session_state.current_page == "Inicio":
-    landing_page()
+
